refactor(query): refactor left outer join to new join
