+++ src/query/service/src/physical_plans/physical_hash_join.rs
@@ -55,9 +55,10 @@ use crate::physical_plans::physical_plan::PhysicalPlan;
 use crate::physical_plans::physical_plan::PhysicalPlanMeta;
 use crate::physical_plans::Exchange;
 use crate::physical_plans::PhysicalPlanBuilder;
-use crate::pipelines::processors::transforms::HashJoinMemoryState;
+use crate::pipelines::processors::transforms::memory::outer_left_join::OuterLeftHashJoin;
+use crate::pipelines::processors::transforms::BasicHashJoinState;
 use crate::pipelines::processors::transforms::HashJoinProbeState;
-use crate::pipelines::processors::transforms::MemoryInnerJoin;
+use crate::pipelines::processors::transforms::InnerHashJoin;
 use crate::pipelines::processors::transforms::PlanRuntimeFilterDesc;
 use crate::pipelines::processors::transforms::TransformHashJoin;
 use crate::pipelines::processors::transforms::TransformHashJoinBuild;
@@ -261,7 +262,7 @@ impl IPhysicalPlan for HashJoin {
         let (enable_optimization, _) = builder.merge_into_get_optimization_flag(self);
 
         if desc.single_to_inner.is_none()
-            && self.join_type == JoinType::Inner
+            && (self.join_type == JoinType::Inner || self.join_type == JoinType::Left)
             && experimental_new_join
             && !enable_optimization
         {
@@ -394,7 +395,7 @@ impl HashJoin {
         builder: &mut PipelineBuilder,
         desc: Arc<HashJoinDesc>,
     ) -> Result<()> {
-        let state = Arc::new(HashJoinMemoryState::create());
+        let state = Arc::new(BasicHashJoinState::create());
         // We must build the runtime filter before constructing the child nodes,
         // as we will inject some runtime filter information into the context for the child nodes to use.
         let rf_desc = PlanRuntimeFilterDesc::create(&builder.ctx, self);
@@ -440,7 +441,7 @@ impl HashJoin {
                 build_input.clone(),
                 probe_input.clone(),
                 joined_output.clone(),
-                self.create_join(builder, desc.clone(), state.clone())?,
+                self.create_join(&self.join_type, builder, desc.clone(), state.clone())?,
                 stage_sync_barrier.clone(),
                 self.projections.clone(),
                 rf_desc.clone(),
@@ -466,10 +467,11 @@ impl HashJoin {
 
     fn create_join(
         &self,
+        join_type: &JoinType,
         builder: &mut PipelineBuilder,
         desc: Arc<HashJoinDesc>,
-        state: Arc<HashJoinMemoryState>,
-    ) -> Result<Box<MemoryInnerJoin>> {
+        state: Arc<BasicHashJoinState>,
+    ) -> Result<Box<dyn crate::pipelines::processors::transforms::Join>> {
         let hash_key_types = self
             .build_keys
             .iter()
@@ -486,13 +488,23 @@ impl HashJoin {
 
         let method = DataBlock::choose_hash_method_with_types(&hash_key_types)?;
 
-        Ok(Box::new(MemoryInnerJoin::create(
-            &builder.ctx,
-            builder.func_ctx.clone(),
-            method,
-            desc,
-            state,
-        )?))
+        Ok(match join_type {
+            JoinType::Inner => Box::new(InnerHashJoin::create(
+                &builder.ctx,
+                builder.func_ctx.clone(),
+                method,
+                desc,
+                state,
+            )?),
+            JoinType::Left => Box::new(OuterLeftHashJoin::create(
+                &builder.ctx,
+                builder.func_ctx.clone(),
+                method,
+                desc,
+                state,
+            )?),
+            _ => unreachable!(),
+        })
     }
 }
